requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
autorootcwd==1.0.1
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
idna==3.11
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
markupsafe==2.1.5
matplotlib==3.10.7
mediapipe==0.10.21
ml-dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==4.25.8
pycparser==2.23
pyparsing==3.2.5
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.5
pytz==2025.2
pyyaml==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
torchmetrics==1.8.2
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
werkzeug==3.1.3
yarl==1.22.0
